Fix logger service validation
This validation cannot be done within the DIC extension, because at
compile time `MergeExtensionConfigurationPass` creates a new
`ContainerBuilder` instance for each extension. Hence, when passing via
configuration a specific service from outside this bundle, this service
cannot be found because it's not created by the extension.

```
[Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException]
  You have requested a non-existent service "my_logger_service".
```

This patch moves the validation logic in a new compiler pass. This
pass is triggered at Optimize step on purpose, in order to have
container parameter placeholders resolved.